Separate plugin from it's symbol state.
Previously a plugin would instal into it's symbol. Now, leave that
symbol empty by default, just allocate the symbol & record it at
the corresponding index.

To do this, we create a `symbols` array that corresponds to `plugins`,
and holds the value of each. We also record the current index of each
plugin in the plug.